AG-16370-fix-grouping-docs
