+++ documentation/ag-grid-docs/src/content/docs/grouping-edit/_examples/refresh-after-group-edit/main.ts
@@ -10,8 +10,27 @@ import {
 } from 'ag-grid-community';
 import { RowGroupingModule } from 'ag-grid-enterprise';
 
-import type { GroupAssignment } from './data';
-import { REGION_LIST, getAssignments } from './data';
+type Region = 'East' | 'West' | 'North' | 'South';
+
+export interface GroupAssignment {
+    id: string;
+    region: Region;
+    owner: string;
+}
+
+function getData(): GroupAssignment[] {
+    return [
+        { id: 'g1', region: 'East', owner: 'Jamie' },
+        { id: 'g2', region: 'East', owner: 'Kira' },
+        { id: 'g3', region: 'East', owner: 'Lena' },
+        { id: 'g4', region: 'West', owner: 'Marco' },
+        { id: 'g5', region: 'West', owner: 'Gus' },
+        { id: 'g6', region: 'North', owner: 'Olive' },
+        { id: 'g7', region: 'North', owner: 'Seth' },
+        { id: 'g8', region: 'South', owner: 'Tara' },
+        { id: 'g9', region: 'South', owner: 'Uma' },
+    ];
+}
 
 ModuleRegistry.registerModules([
     RowGroupingModule,
@@ -30,7 +49,7 @@ const gridOptions: GridOptions<GroupAssignment> = {
             rowGroup: true,
             editable: true,
             cellEditor: 'agSelectCellEditor',
-            cellEditorParams: { values: REGION_LIST },
+            cellEditorParams: { values: ['East', 'West', 'North', 'South'] },
         },
         { field: 'owner' },
     ],
@@ -42,7 +61,7 @@ const gridOptions: GridOptions<GroupAssignment> = {
     autoGroupColumnDef: {
         minWidth: 220,
     },
-    rowData: getAssignments(),
+    rowData: getData(),
     refreshAfterGroupEdit: true,
     groupDefaultExpanded: -1,
     animateRows: true,
