+++ documentation/ag-grid-docs/src/content/docs/grouping-row-dragging/_examples/managed-row-group-drag-multi-level/data.ts
@@ -1,16 +1,4 @@
-export interface IAthlete {
-    id: string;
-    athlete: string;
-    age: number | null;
-    country: string | null;
-    year: number;
-    date: string | null;
-    sport: string;
-    gold: number;
-    silver: number;
-    bronze: number;
-    total: number;
-}
+import type { IAthlete } from './types';
 
 export const getAthletesData = (): IAthlete[] => {
     return [
