+++ documentation/ag-grid-docs/src/content/docs/grouping-row-dragging/_examples/managed-row-group-drag-multi-level/types.ts
@@ -1,0 +1,13 @@
+export interface IAthlete {
+    id: string;
+    athlete: string;
+    age: number | null;
+    country: string | null;
+    year: number;
+    date: string | null;
+    sport: string;
+    gold: number;
+    silver: number;
+    bronze: number;
+    total: number;
+}
